Use the inotify API on linux to improve performance
Now the gamepad extension only iterates over each input device when initialized
or when there's an error while reading the inotify events. Any new input device
is queried whether it's a gamepad and added if it is one every frame, so we
still have plug-n-play equivalent to Windows and Emscripten.